+++ NEWS.md
@@ -1,5 +1,12 @@
+1.3.0 (DD/MM/YY)
+-----
+* Support for specifying a filename for exports (#64)
+* Support for specifying a list of fields to skip from output (#63)
+* Add 'cache_policy' parameter (#62)
+
 1.2.1 (DD/MM/YY)
 -----
+* Added timeout default to 600 miliseconds in cluster.js
 
 1.2.0 (DD/MM/YY)
 -----